Deactivate nohlsearch when exiting command line without <CR>
Previously I was deactivating the <CR> mapping in the rhs of the <CR>
mapping. However, you can also exit command mode without pressing <CR>
(e.g., with <C-c>). In that case, the <CR> mapping would still be
active, causing regular search commands to nohlsearch on completion.

Now, use a `CmdlineLeave` `autocmd` instead to remove the mapping and
autocmd. This ensures the plugin always cleans up after itself when the
search line is exited.